добавили удаление записей

+++ http_post_edit.go
@@ -109,7 +109,7 @@ func http_post_edit__load_data_fn(w http.ResponseWriter, r *http.Request, d map[
 			  UNION ALL
 			  SELECT FIRST 1 SKIP 0 
 	            1 AS n,p.name,p.tags,p.preview,p.text,p.uuid_user,LEFT(p.date_create,16) AS date_create,p.uuid,p.edit_type,p.uuid_user_create,p.uuid_user_publish,LEFT(p.date_modify,16) AS date_modify
-	          FROM tpost p WHERE p.uuid='` + id + `' AND p.uuid_user='` + uuid_user + `'
+	          FROM tpost p WHERE p.uuid='` + id + `' AND p.uuid_user='` + uuid_user + `' AND p.edit_type!='publish'
 			  )
 			  ORDER BY n
 			`
@@ -183,6 +183,12 @@ func http_post_edit_ajax(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if rtype == "delete" {
+		http_post_edit__delete_data(w, r, d)
+		RenderJson(w, r, d)
+		return
+	}
+
 	if rtype == "publish" {
 		http_post_edit__publish(w, r, d)
 		RenderJson(w, r, d)
@@ -194,7 +200,6 @@ func http_post_edit_ajax(w http.ResponseWriter, r *http.Request) {
 	RenderJson(w, r, d)
 }
 
-//сохраняем данные поста в таблицу tpost_edit
 //должены быть заданы post данные
 //  d["data"]["uuid"] - uuid поста если все успешно
 //  d["error"] - если ошибки
@@ -273,24 +278,119 @@ func http_post_edit__save_data(w http.ResponseWriter, r *http.Request, d map[str
 		//если пользователь первый раз редактирует эту запись то создаем его версию этой записи
 		data["edit_type"] = "update"
 		query = `INSERT INTO TPOST(NAME,TAGS,TEXT,PREVIEW,UUID_USER,
-				   DATE_CREATE,
+				   DATE_CREATE,uuid_user_create,
 				   UUID/*,UUID_USER_PUBLISH*/,EDIT_TYPE
 				   )
 			    VALUES('` + name + `','` + tags + `','` + text + `','` + preview + `','` + uuid_user_this + `',
-				   (SELECT MIN(date_create) FROM tpost WHERE uuid='` + uuid + `'),
+				   (SELECT MIN(date_create) FROM tpost WHERE uuid='` + uuid + `'),(SELECT MIN(uuid_user_create) FROM tpost WHERE uuid='` + uuid + `'),
 				   '` + uuid + `'/*,'` + uuid_user_this + `'*/,'` + data["edit_type"].(string) + `'
 				   )`
 	} else {
 		//если пользователь создает новую запись то создаем его версию этой записи
 		data["edit_type"] = "create"
 		query = `INSERT INTO TPOST(NAME,TAGS,TEXT,PREVIEW,UUID_USER,
-				   DATE_CREATE,
+				   DATE_CREATE,uuid_user_create,
 				   UUID/*,UUID_USER_PUBLISH*/,EDIT_TYPE,UUID_USER_CREATE
 				   )
 			    VALUES('` + name + `','` + tags + `','` + text + `','` + preview + `','` + uuid_user_this + `',
-				   (SELECT MIN(date_create) FROM tpost WHERE uuid='` + uuid + `'),
+				   (SELECT MIN(date_create) FROM tpost WHERE uuid='` + uuid + `'),(SELECT MIN(uuid_user_create) FROM tpost WHERE uuid='` + uuid + `'),
 				   '` + uuid + `'/*,'` + uuid_user_this + `'*/,'` + data["edit_type"].(string) + `','` + uuid_user_this + `'
 				   )`
+	}
+	//LogPrint("=====================================================================")
+	//LogPrint(fmt.Sprintf("edit_type: %#v", data["edit_type"]))
+	//LogPrint(fmt.Sprintf("%#v", query))
+	//LogPrint("=====================================================================")
+	run_db_exec(d, query)
+
+	http_post_edit__load_data_fn(w, r, d, uuid, uuid_user_this)
+
+	return
+}
+
+//должены быть заданы post данные
+//  d["data"]["uuid"] - uuid поста если все успешно
+//  d["error"] - если ошибки
+func http_post_edit__delete_data(w http.ResponseWriter, r *http.Request, d map[string]interface{}) {
+	r.ParseForm()
+	//LogPrint("=====================================================================")
+	//LogPrint(fmt.Sprintf("%#v", r.Form))
+	//LogPrint("=====================================================================")
+
+	uuid_user_this := get_map_val(d, "", "user", "uuid_user").(string)
+	uuid_user := http_get_var_str(r.Form, "uuid_user", "")
+	if uuid_user == "" {
+		uuid_user = uuid_user_this
+	}
+
+	if uuid_user != uuid_user_this { //если юзер пытается загрузить для редактирования чужую запись
+		if d["db_access"].(string) < "3" { // то у него должны быть права модератора
+			d["error"] = fmt.Errorf("%s", "у вас нет доступа к редактированию чужих записей в БД \""+d["db"].(*DBd).Name+"\"")
+			d["errorcode"] = "dbnoaccess"
+			return
+		}
+	}
+	uuid_user = strings.Replace(uuid_user, "'", "''", -1)
+
+	uuid := http_get_var_str_sql(r.Form, "uuid", "")
+	if uuid == "" {
+		d["error"] = fmt.Errorf("%s", "не верно указан/не указан POST параметр \"uuid\":\""+uuid+"\"")
+		d["errorcode"] = "nopostparam"
+		return
+	}
+
+	//проверяем есть ли обновляемые записи в бд
+	query := `SELECT 
+				(SELECT COUNT(*) FROM tpost p WHERE p.uuid='` + uuid + `') AS c1,
+				(SELECT COUNT(*) FROM tpost p WHERE p.uuid='` + uuid + `' AND p.uuid_user='` + uuid_user + `' AND p.edit_type!='publish') AS c2
+			  FROM rdb$database`
+	rows := run_db_query(d, query)
+	if rows == nil {
+		return
+	}
+	var c1, c2 int
+	for rows.Next() {
+		if err := rows.Scan(&c1, &c2); err != nil {
+			d["error"] = fmtError("http_post_edit__delete_data ERROR001 rows.Scan: query:\n"+query+"\n\n", err)
+			d["errorcode"] = "dbqueryerror"
+			return
+		}
+	}
+	rows.Close()
+
+	data := make(map[string]interface{})
+	d["data"] = data
+	name := http_get_var_str_sql(r.Form, "name", "")
+	text := http_get_var_str_sql(r.Form, "text", "")
+	tags := http_get_var_str_sql(r.Form, "tags", "")
+	preview := http_get_var_str_sql(r.Form, "preview", "")
+
+	query = "hz che delatb nax"
+	if c2 > 0 {
+		//если пользователь ранее редактировал эту запись и она не публиковалась то просто обновляем его версию его записи
+		data["edit_type"] = "delete"
+		query = `UPDATE TPOST p SET 
+					UUID_USER='` + uuid_user_this + `',
+					EDIT_TYPE='` + data["edit_type"].(string) + `'
+			    WHERE p.uuid='` + uuid + `' 
+				  AND p.uuid_user='` + uuid_user + `' 
+				  AND p.edit_type!='publish'`
+	} else if c1 > 0 {
+		//если пользователь первый раз редактирует эту запись то создаем его версию этой записи
+		data["edit_type"] = "delete"
+		query = `INSERT INTO TPOST(NAME,TAGS,TEXT,PREVIEW,UUID_USER,
+				   DATE_CREATE,uuid_user_create,
+				   UUID/*,UUID_USER_PUBLISH*/,EDIT_TYPE
+				   )
+			    VALUES('` + name + `','` + tags + `','` + text + `','` + preview + `','` + uuid_user_this + `',
+				   (SELECT MIN(date_create) FROM tpost WHERE uuid='` + uuid + `'),(SELECT MIN(uuid_user_create) FROM tpost WHERE uuid='` + uuid + `'),
+				   '` + uuid + `'/*,'` + uuid_user_this + `'*/,'` + data["edit_type"].(string) + `'
+				   )`
+	} else {
+		//если запись не существует
+		d["error"] = ("запись не найдена (возможно она уже удалена): uuid: " + uuid + " / uuid_user: " + uuid_user + " ")
+		d["errorcode"] = "recordnotfound"
+		return
 	}
 	//LogPrint("=====================================================================")
 	//LogPrint(fmt.Sprintf("edit_type: %#v", data["edit_type"]))
@@ -335,15 +435,16 @@ func http_post_edit__publish(w http.ResponseWriter, r *http.Request, d map[strin
 	//проверяем есть ли обновляемые записи в бд
 	query := `SELECT 
 				(SELECT COUNT(*) FROM tpost p WHERE p.uuid='` + uuid + `' AND p.edit_type='publish') AS c1,
-				(SELECT COUNT(*) FROM tpost p WHERE p.uuid='` + uuid + `' AND p.uuid_user='` + uuid_user + `' AND p.edit_type!='publish') AS c2
+				(SELECT COUNT(*) FROM tpost p WHERE p.uuid='` + uuid + `' AND p.uuid_user='` + uuid_user + `' AND p.edit_type!='publish') AS c2,
+				(SELECT COUNT(*) FROM tpost p WHERE p.uuid='` + uuid + `' AND p.uuid_user='` + uuid_user + `' AND p.edit_type='delete') AS c3
 			  FROM rdb$database`
 	rows := run_db_query(d, query)
 	if rows == nil {
 		return
 	}
-	var c1, c2 int
+	var c1, c2, c3 int
 	for rows.Next() {
-		if err := rows.Scan(&c1, &c2); err != nil {
+		if err := rows.Scan(&c1, &c2, &c3); err != nil {
 			d["error"] = fmtError("http_post_edit__publish ERROR001 rows.Scan: query:\n"+query+"\n\n", err)
 			d["errorcode"] = "dbqueryerror"
 			return
@@ -354,12 +455,13 @@ func http_post_edit__publish(w http.ResponseWriter, r *http.Request, d map[strin
 	data := make(map[string]interface{})
 	d["data"] = data
 
-	if c2 > 0 && c1 > 0 {
-		//если пост существует и существует ранее не публиковавшаяся отредактированная запись этого поста указанным юзером
-		//то обновляем текущий опубликованный пост из отредактированной записи и удаляем её за ненадобностью
+	if c3 == 0 {
+		if c2 > 0 && c1 > 0 {
+			//если пост существует и существует ранее не публиковавшаяся отредактированная запись этого поста указанным юзером
+			//то обновляем текущий опубликованный пост из отредактированной записи и удаляем её за ненадобностью
 
-		//обновляем существующую опубликованную запись (именно не удаляем а меняем значения полей - это надо для логов)
-		query = `UPDATE tpost p SET 
+			//обновляем существующую опубликованную запись (именно не удаляем а меняем значения полей - это надо для логов)
+			query = `UPDATE tpost p SET 
 					p.name =         (SELECT t.name FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
 					p.text =         (SELECT t.text FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
 					p.tags =         (SELECT t.tags FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
@@ -375,31 +477,31 @@ func http_post_edit__publish(w http.ResponseWriter, r *http.Request, d map[strin
 			    WHERE p.uuid='` + uuid + `' 
 				  AND p.edit_type='publish'
 				`
-		b := run_db_exec(d, query)
-		if b == nil {
-			return
-		}
+			b := run_db_exec(d, query)
+			if b == nil {
+				return
+			}
 
-		//удаляем запись которую пользователь редактировал (она больше не нужна)
-		query = `DELETE FROM tpost p 
+			//удаляем запись которую пользователь редактировал (она больше не нужна)
+			query = `DELETE FROM tpost p 
 			    WHERE p.uuid='` + uuid + `' 
 				  AND p.uuid_user='` + uuid_user + `' 
 				  AND p.edit_type!='publish'`
-		b = run_db_exec(d, query)
-		if b == nil {
+			b = run_db_exec(d, query)
+			if b == nil {
+				return
+			}
+
+			LogPrint(fmt.Sprintf("publish update record id: %s; user: %s", uuid, uuid_user))
+
+		} else if c1 > 0 && c2 == 0 {
+			//если пользователь пытается опубликовать запись но до этого её не редактировал (не сохранял изменения)
+			d["error"] = "нет обновленных(сохраненных) данных для публикации записи"
+			d["errorcode"] = "nochanges"
 			return
-		}
-
-		LogPrint(fmt.Sprintf("publish update record id: %s; user: %s", uuid, uuid_user))
-
-	} else if c1 > 0 && c2 == 0 {
-		//если пользователь пытается опубликовать запись но до этого её не редактировал (не сохранял изменения)
-		d["error"] = "нет обновленных(сохраненных) данных для публикации записи"
-		d["errorcode"] = "nochanges"
-		return
-	} else if c1 == 0 && c2 > 0 {
-		//если пользователь создает новую запись и до этого она не публиковалась или была удалена
-		query = `INSERT INTO TPOST(NAME,TAGS,TEXT,PREVIEW,UUID_USER,
+		} else if c1 == 0 && c2 > 0 {
+			//если пользователь создает новую запись и до этого она не публиковалась или была удалена
+			query = `INSERT INTO TPOST(NAME,TAGS,TEXT,PREVIEW,UUID_USER,
 				   		DATE_CREATE,date_modify,uuid_user_create,
 				   		UUID,UUID_USER_PUBLISH,EDIT_TYPE,date_publish
 				    )
@@ -409,21 +511,68 @@ func http_post_edit__publish(w http.ResponseWriter, r *http.Request, d map[strin
 				 FROM tpost 
 				 WHERE uuid='` + uuid + `' AND uuid_user='` + uuid_user + `'
 				`
-		b := run_db_exec(d, query)
-		if b == nil {
-			return
-		}
-		//удаляем запись которую пользователь редактировал (она больше не нужна)
-		query = `DELETE FROM tpost p 
+			b := run_db_exec(d, query)
+			if b == nil {
+				return
+			}
+			//удаляем запись которую пользователь редактировал (она больше не нужна)
+			query = `DELETE FROM tpost p 
 			    WHERE p.uuid='` + uuid + `' 
 				  AND p.uuid_user='` + uuid_user + `' 
 				  AND p.edit_type!='publish'
 				`
-		b = run_db_exec(d, query)
-		if b == nil {
+			b = run_db_exec(d, query)
+			if b == nil {
+				return
+			}
+			LogPrint(fmt.Sprintf("publish new record id: %s; user: %s", uuid, uuid_user))
+		}
+	} else if c3 > 0 { //если удаляем запись
+		if c1 > 0 {
+			//если пост существует то удаляем текущий опубликованный пост
+
+			//обновляем существующую опубликованную запись (именно не удаляем а меняем значения полей - это надо для логов)
+			query = `UPDATE tpost p SET 
+					p.name =         (SELECT t.name FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
+					p.text =         (SELECT t.text FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
+					p.tags =         (SELECT t.tags FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
+					p.preview =      (SELECT t.preview FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
+					p.uuid_user =    (SELECT t.uuid_user FROM tpost t WHERE t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'),
+					
+					p.date_modify   = current_timestamp,
+					p.date_publish  = current_timestamp,
+					p.uuid_user_publish = '` + uuid_user_this + `',
+					p.edit_type = 'delete'
+				--FROM tpost p 
+				--  LEFT JOIN tpost t ON t.uuid=p.uuid AND t.uuid_user='` + uuid_user + `' AND t.edit_type!='publish'
+			    WHERE p.uuid='` + uuid + `' 
+				  AND p.edit_type='publish'
+				`
+			b := run_db_exec(d, query)
+			if b == nil {
+				return
+			}
+
+			//удаляем опубликованную запись
+			query = `DELETE FROM tpost p 
+			    WHERE p.uuid='` + uuid + `'
+				  AND p.uuid_user = '` + uuid_user + `'
+				  AND p.uuid_user_publish IS NULL
+				  AND p.date_publish IS NULL
+				  AND p.edit_type='delete'`
+			b = run_db_exec(d, query)
+			if b == nil {
+				return
+			}
+
+			LogPrint(fmt.Sprintf("delete record id: %s; user: %s", uuid, uuid_user))
+
+		} else if c1 == 0 {
+			//если пользователь пытается удалить запись которая не публиковалась
+			d["error"] = "указанная запись не опубликована (её не удалить из опубликованных записей)"
+			d["errorcode"] = "nochanges"
 			return
 		}
-		LogPrint(fmt.Sprintf("publish new record id: %s; user: %s", uuid, uuid_user))
 	}
 	//LogPrint("=====================================================================")
 	//LogPrint(fmt.Sprintf("edit_type: %#v", data["edit_type"]))
